# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Üs Koruma Karar Destek",
    "version": "0.1.0"
  },
  "paths": {
...
    "/images/{image_id}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Image",
        "description": "Görüntünün boyutu, çekim saati, köşe koordinatları, merkezi ve bölgesi.",
        "operationId": "get_image_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDetail"
                }
              }
            }
          },
          "404": {
            "description": "Görüntü veri setinde yok"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/{image_id}/file": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Image File",
        "description": "Görüntü dosyası: Supabase modunda `images.file_path` ile bucket'tan, ağsız demoda\nyerel klasörden. Yol veri setindeki kayıttan kurulur, istekten gelen yol kullanılmaz.",
        "operationId": "get_image_file_images__image_id__file_get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Görüntü dosyası",
            "content": {
              "image/jpeg": {},
              "image/png": {}
            }
          },
          "404": {
            "description": "Görüntü veri setinde yok ya da dosyası bulunamadı"
          },
          "502": {
            "description": "Storage'a ulaşılamadı"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CornerCoordinates": {
        "properties": {
          "top_left": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Top Left"
          },
          "top_right": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Top Right"
          },
          "bottom_left": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Bottom Left"
          },
          "bottom_right": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Bottom Right"
          }
        },
        "type": "object",
        "required": [
          "top_left",
          "top_right",
          "bottom_left",
          "bottom_right"
        ],
        "title": "CornerCoordinates"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageDetail": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "width_px": {
            "type": "integer",
            "title": "Width Px"
          },
          "height_px": {
            "type": "integer",
            "title": "Height Px"
          },
          "capture_time": {
            "type": "string",
            "title": "Capture Time"
          },
          "corner_coordinates": {
            "$ref": "#/components/schemas/CornerCoordinates"
          },
          "center": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Center"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "width_px",
          "height_px",
          "capture_time",
          "corner_coordinates",
          "center",
          "zone"
        ],
        "title": "ImageDetail",
        "description": "Görüntünün `image_meta.json` kaydı; karenin merkezi ve bölgesiyle."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}